refactor(evolution): fold cases into audit module
